+++ tests/PSRule.Tests/ObjectHelperTests.cs
@@ -17,18 +17,21 @@ namespace PSRule
             Runtime.ObjectHelper.GetField(bindingContext: null, targetObject: testObject, name: "Value.Value1", caseSensitive: false, value: out object actual2);
             Runtime.ObjectHelper.GetField(bindingContext: null, targetObject: testObject, name: "Metadata.'app.kubernetes.io/name'", caseSensitive: false, value: out object actual3);
             Runtime.ObjectHelper.GetField(bindingContext: null, targetObject: testObject, name: "Value2[1]", caseSensitive: false, value: out object actual4);
+            Runtime.ObjectHelper.GetField(bindingContext: null, targetObject: testObject, name: ".", caseSensitive: true, value: out object actual5);
+            Runtime.ObjectHelper.GetField(bindingContext: null, targetObject: testObject, name: ".Value2[1]", caseSensitive: false, value: out object actual6);
 
             Assert.Equal(expected: testObject.Name, actual: actual1);
             Assert.Equal(expected: testObject.Value.Value1, actual: actual2);
             Assert.Equal(expected: testObject.Metadata["app.kubernetes.io/name"], actual: actual3);
             Assert.Equal(expected: testObject.Value2[1], actual: actual4);
+            Assert.Equal(expected: testObject, actual: actual5);
+            Assert.Equal(expected: testObject.Value2[1], actual: actual6);
         }
 
-        private TestObject1 GetTestObject()
+        private static TestObject1 GetTestObject()
         {
             var result = new TestObject1 { Name = "TestObject1", Value = new TestObject2 { Value1 = "Value1" }, Value2 = new string[] { "1", "2" }, Metadata = new Hashtable() };
             result.Metadata.Add("app.kubernetes.io/name", "KubeName");
-
             return result;
         }
 
@@ -47,6 +50,5 @@ namespace PSRule
         {
             public string Value1;
         }
-
     }
 }